[AI] Revert rename of `GeminiModel` to `GeminiLanguageModel` (#16227)
