+++ FirebaseAI/Sources/FirebaseAI.swift
@@ -109,7 +109,7 @@ public final class FirebaseAI: Sendable {
   // TODO: Remove the `#if compiler(>=6.2.3)` when Xcode 26.2 is the minimum supported version.
   #if compiler(>=6.2.3)
 
-    /// **[Public Preview]** Creates a ``GeminiLanguageModel`` for the specified model name.
+    /// **[Public Preview]** Creates a ``GeminiModel`` for the specified model name.
     ///
     /// > Warning: This API is a public preview and may be subject to change.
     ///
@@ -120,10 +120,10 @@ public final class FirebaseAI: Sendable {
     ///   - safetySettings: Settings to adjust the safety filters for potentially harmful content
     ///     generated by the model.
     ///   - requestOptions: Configuration parameters for sending requests to the backend.
-    public func geminiLanguageModel(name: String, safetySettings: [SafetySetting]? = nil,
-                                    requestOptions: RequestOptions = RequestOptions())
-      -> GeminiLanguageModel {
-      return GeminiLanguageModel(
+    public func geminiModel(name: String, safetySettings: [SafetySetting]? = nil,
+                            requestOptions: RequestOptions = RequestOptions())
+      -> GeminiModel {
+      return GeminiModel(
         modelName: name,
         modelResourceName: modelResourceName(modelName: name),
         firebaseInfo: firebaseInfo,
