+++ FirebaseAI/Sources/Protocols/Public/LanguageModelProvider.swift
@@ -36,8 +36,8 @@
     ///   - safetySettings: Settings to adjust the safety filters for potentially harmful content
     ///     generated by the model.
     ///   - requestOptions: Configuration parameters for sending requests to the backend.
-    static func geminiLanguageModel(name: String, safetySettings: [SafetySetting]? = nil,
-                                    requestOptions: RequestOptions = RequestOptions())
+    static func geminiModel(name: String, safetySettings: [SafetySetting]? = nil,
+                            requestOptions: RequestOptions = RequestOptions())
       -> GeminiModelProvider {
       return GeminiModelProvider(
         modelName: name,
@@ -88,7 +88,7 @@
   // simple `String` containing a Gemini model name.
   extension String: LanguageModelProvider {
     public func _languageModel(firebaseAI: FirebaseAI) -> any FirebaseAI.LanguageModel {
-      return firebaseAI.geminiLanguageModel(name: self)
+      return firebaseAI.geminiModel(name: self)
     }
   }
 #endif // compiler(>=6.2.3)
