--- FirebaseAI/Sources/Types/Public/GeminiLanguageModel.swift
@@ -1,137 +1,0 @@
-// Copyright 2026 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-import FirebaseCore
-import Foundation
-
-#if compiler(>=6.2.3)
-  /// A Gemini model accessed via the Gemini Developer API or the Vertex AI API.
-  ///
-  /// **Public Preview**: This API is a public preview and may be subject to change.
-  ///
-  /// To create a ``GeminiLanguageModel`` instance, see
-  /// ``FirebaseAI/geminiLanguageModel(name:safetySettings:requestOptions:)``.
-  public struct GeminiLanguageModel {
-    let modelConfig: ModelConfig
-    let modelResourceName: String
-    let firebaseInfo: FirebaseInfo
-    let toolConfig: ToolConfig?
-    let urlSession: URLSession
-
-    init(modelName: String,
-         modelResourceName: String,
-         firebaseInfo: FirebaseInfo,
-         apiConfig: APIConfig,
-         safetySettings: [SafetySetting]? = nil,
-         toolConfig: ToolConfig? = nil,
-         requestOptions: RequestOptions = RequestOptions(),
-         urlSession: URLSession = GenAIURLSession.default) {
-      modelConfig = ModelConfig(
-        modelName: modelName,
-        appName: firebaseInfo.app.name,
-        apiConfig: apiConfig,
-        useLimitedUseAppCheckTokens: firebaseInfo.useLimitedUseAppCheckTokens,
-        safetySettings: safetySettings,
-        requestOptions: requestOptions
-      )
-      self.modelResourceName = modelResourceName
-      self.firebaseInfo = firebaseInfo
-      self.toolConfig = toolConfig
-      self.urlSession = urlSession
-    }
-
-    /// **[Public Preview]** Creates a ``GeminiLanguageModel`` for the specified model name.
-    ///
-    /// > Warning: This API is a public preview and may be subject to change.
-    ///
-    /// - Parameters:
-    ///   - name: The name of the model to use; see [available model names
-    ///     ](https://firebase.google.com/docs/vertex-ai/gemini-models#available-model-names)
-    ///     for a list of supported model names.
-    ///   - firebaseAI: A ``FirebaseAI`` instance that provides the Firebase app and backend
-    ///     configuration details for the Gemini model. If not specified, this is configured with
-    ///     the default `FirebaseApp` and the ``Backend/googleAI()`` (Gemini Developer API) backend.
-    ///     To use the Vertex AI backend, specify
-    ///     `.firebaseAI(backend: .vertexAI(location: "global"))`; see
-    ///     ``FirebaseAI/firebaseAI(app:backend:useLimitedUseAppCheckTokens:)`` and
-    ///     ``Backend/vertexAI(location:)`` for more configuration options and details.
-    ///   - safetySettings: Settings to adjust the safety filters for potentially harmful content
-    ///     generated by the model.
-    ///   - requestOptions: Configuration parameters for sending requests to the backend.
-    public init(name: String, firebaseAI: FirebaseAI = FirebaseAI.firebaseAI(),
-                safetySettings: [SafetySetting]? = nil,
-                requestOptions: RequestOptions = RequestOptions()) {
-      self.init(
-        modelName: name,
-        modelResourceName: firebaseAI.modelResourceName(modelName: name),
-        firebaseInfo: firebaseAI.firebaseInfo,
-        apiConfig: firebaseAI.apiConfig,
-        safetySettings: safetySettings,
-        requestOptions: requestOptions,
-      )
-    }
-
-    struct ModelConfig: Sendable, Hashable {
-      let modelName: String
-      let appName: String?
-      let apiConfig: APIConfig
-      let useLimitedUseAppCheckTokens: Bool
-      let safetySettings: [SafetySetting]?
-      let requestOptions: RequestOptions
-
-      var firebaseApp: FirebaseApp? {
-        return appName.flatMap { FirebaseApp.app(name: $0) }
-      }
-
-      var firebaseAI: FirebaseAI {
-        return FirebaseAI.createInstance(
-          app: firebaseApp,
-          apiConfig: apiConfig,
-          useLimitedUseAppCheckTokens: useLimitedUseAppCheckTokens
-        )
-      }
-    }
-  }
-
-  extension GeminiLanguageModel: FirebaseAI.LanguageModel {
-    /// Returns the name of the model.
-    ///
-    /// > Important: This property is for **internal use only** and may change at any time.
-    public var _modelName: String {
-      return modelConfig.modelName
-    }
-
-    /// Returns a new session for this model.
-    ///
-    /// > Important: This method is for **internal use only** and may change at any time.
-    public func _startSession(tools: [any ToolRepresentable]?,
-                              instructions: String?) throws -> any _ModelSession {
-      let model = GenerativeModel(
-        modelName: modelConfig.modelName,
-        modelResourceName: modelResourceName,
-        firebaseInfo: firebaseInfo,
-        apiConfig: modelConfig.apiConfig,
-        generationConfig: nil,
-        safetySettings: modelConfig.safetySettings,
-        tools: tools?.map { $0.toolRepresentation },
-        // TODO: Add toolConfig
-        systemInstruction: instructions.map { ModelContent(role: "system", parts: $0) },
-        requestOptions: modelConfig.requestOptions,
-        urlSession: urlSession
-      )
-
-      return GeminiModelSession(model: model, history: [])
-    }
-  }
-#endif // compiler(>=6.2.3)
