+++ FirebaseAI/Tests/TestApp/Tests/Integration/GenerativeModelSessionHybridTests.swift
@@ -27,12 +27,9 @@
     @Test(arguments: [InstanceConfig.vertexAI_v1beta_global])
     func respondText_fallbackOnGeminiModelError(_ config: InstanceConfig) async throws {
       let firebaseAI = FirebaseAI.componentInstance(config)
-      let invalidModel1 = GeminiLanguageModel(name: "invalid-model-name-1")
-      let invalidModel2 = GeminiLanguageModel(name: "invalid-model-name-2")
-      let validModel = GeminiLanguageModel(
-        name: ModelNames.gemini2_5_FlashLite,
-        firebaseAI: firebaseAI
-      )
+      let invalidModel1 = firebaseAI.geminiModel(name: "invalid-model-name-1")
+      let invalidModel2 = firebaseAI.geminiModel(name: "invalid-model-name-2")
+      let validModel = firebaseAI.geminiModel(name: ModelNames.gemini2_5_FlashLite)
       let session = firebaseAI.generativeModelSession(
         model: .hybridModel(
           primary: invalidModel1,
@@ -66,12 +63,9 @@
     func respondText_fallbackOnFoundationModelsError(_ config: InstanceConfig) async throws {
       let firebaseAI = FirebaseAI.componentInstance(config)
       let systemModel = FirebaseAI.SystemLanguageModel.default
-      let geminiLanguageModel = GeminiLanguageModel(
-        name: ModelNames.gemini2_5_FlashLite,
-        firebaseAI: firebaseAI
-      )
+      let geminiModel = firebaseAI.geminiModel(name: ModelNames.gemini2_5_FlashLite)
       let session = firebaseAI.generativeModelSession(
-        model: HybridModel(primary: systemModel, secondary: geminiLanguageModel)
+        model: HybridModel(primary: systemModel, secondary: geminiModel)
       )
       let prompt = "In one sentence, why is the sky blue?"
 
@@ -89,7 +83,7 @@
       if await foundationModelsIsAvailable() {
         #expect(response.rawResponse.modelVersion == systemModel._modelName)
       } else {
-        #expect(response.rawResponse.modelVersion == geminiLanguageModel._modelName)
+        #expect(response.rawResponse.modelVersion == geminiModel._modelName)
       }
     }
 
@@ -99,8 +93,8 @@
     @available(watchOS, unavailable)
     func respondGenerable_fallbackOnGeminiModelError(_ config: InstanceConfig) async throws {
       let firebaseAI = FirebaseAI.componentInstance(config)
-      let invalidModel = GeminiLanguageModel(name: "invalid-model-name-1")
-      let validModel = GeminiLanguageModel(name: ModelNames.gemini2_5_FlashLite)
+      let invalidModel = firebaseAI.geminiModel(name: "invalid-model-name-1")
+      let validModel = firebaseAI.geminiModel(name: ModelNames.gemini2_5_FlashLite)
       let session = firebaseAI.generativeModelSession(
         model: HybridModel(primary: invalidModel, secondary: validModel)
       )
@@ -124,8 +118,8 @@
     @Test(arguments: [InstanceConfig.vertexAI_v1beta_global])
     func streamResponseText_fallbackOnGeminiModelError(_ config: InstanceConfig) async throws {
       let firebaseAI = FirebaseAI.componentInstance(config)
-      let invalidModel = GeminiLanguageModel(name: "invalid-model-name")
-      let validModel = GeminiLanguageModel(name: ModelNames.gemini2_5_FlashLite)
+      let invalidModel = firebaseAI.geminiModel(name: "invalid-model-name")
+      let validModel = firebaseAI.geminiModel(name: ModelNames.gemini2_5_FlashLite)
       let session = firebaseAI.generativeModelSession(
         model: HybridModel(primary: invalidModel, secondary: validModel)
       )
@@ -174,11 +168,8 @@
     @available(watchOS, unavailable)
     func streamResponseGenerable_fallbackOnGeminiModelError(_ config: InstanceConfig) async throws {
       let firebaseAI = FirebaseAI.componentInstance(config)
-      let invalidModel = GeminiLanguageModel(name: "invalid-model-name")
-      let validModel = GeminiLanguageModel(
-        name: ModelNames.gemini2_5_FlashLite,
-        firebaseAI: firebaseAI
-      )
+      let invalidModel = firebaseAI.geminiModel(name: "invalid-model-name")
+      let validModel = firebaseAI.geminiModel(name: ModelNames.gemini2_5_FlashLite)
       let session = firebaseAI.generativeModelSession(
         model: .hybridModel(primary: invalidModel, secondary: validModel)
       )
@@ -224,12 +215,9 @@
     func streamResponseText_fallbackOnFoundationModelsError(_ config: InstanceConfig) async throws {
       let firebaseAI = FirebaseAI.componentInstance(config)
       let systemModel = FirebaseAI.SystemLanguageModel.default
-      let geminiLanguageModel = GeminiLanguageModel(
-        name: ModelNames.gemini2_5_FlashLite,
-        firebaseAI: firebaseAI
-      )
+      let geminiModel = firebaseAI.geminiModel(name: ModelNames.gemini2_5_FlashLite)
       let session = firebaseAI.generativeModelSession(
-        model: HybridModel(primary: systemModel, secondary: geminiLanguageModel)
+        model: HybridModel(primary: systemModel, secondary: geminiModel)
       )
       let prompt = "In one sentence, why is the sky blue?"
 
@@ -251,7 +239,7 @@
       if await foundationModelsIsAvailable() {
         #expect(response.rawResponse.modelVersion == systemModel._modelName)
       } else {
-        #expect(response.rawResponse.modelVersion == geminiLanguageModel._modelName)
+        #expect(response.rawResponse.modelVersion == geminiModel._modelName)
       }
     }
