+++ FirebaseAI/Tests/Unit/GenerativeModelSessionAPITests.swift
@@ -25,21 +25,12 @@ import XCTest
       _ = ai.generativeModelSession(model: "gemini-flash-latest")
 
       // Initialize a session with a `GeminiModelProvider`
-      _ = ai.generativeModelSession(model: .geminiLanguageModel(name: "gemini-flash-latest"))
+      _ = ai.generativeModelSession(model: .geminiModel(name: "gemini-flash-latest"))
 
-      // Initialize a `GeminiLanguageModel`
-      let geminiLanguageModel = ai.geminiLanguageModel(name: "gemini-flash-latest")
-      _ = GeminiLanguageModel(name: "gemini-flash-latest")
-      _ = GeminiLanguageModel(name: "gemini-flash-latest", firebaseAI: ai)
-      _ = GeminiLanguageModel(
-        name: "gemini-flash-latest",
-        firebaseAI: .firebaseAI(backend: .vertexAI(location: "global")),
-        safetySettings: [
-          SafetySetting(harmCategory: .dangerousContent, threshold: .blockLowAndAbove),
-        ]
-      )
+      // Initialize a `GeminiModel`
+      let geminiLanguageModel = ai.geminiModel(name: "gemini-flash-latest")
 
-      // Initialize a session with a `GeminiLanguageModel`
+      // Initialize a session with a `GeminiModel`
       _ = ai.generativeModelSession(model: geminiLanguageModel)
 
       // Initialize a session with a `SystemLanguageModel` as a `FirebaseAI.LanguageModel`
@@ -53,7 +44,7 @@ import XCTest
       _ = ai.generativeModelSession(
         model: .hybridModel(
           primary: geminiLanguageModel,
-          secondary: .geminiLanguageModel(name: "gemini-flash-lite-latest")
+          secondary: .geminiModel(name: "gemini-flash-lite-latest")
         )
       )
 
@@ -61,18 +52,18 @@ import XCTest
       _ = ai.generativeModelSession(
         model: .hybridModel(
           primary: .systemModel(),
-          secondary: .geminiLanguageModel(name: "gemini-flash-lite-latest")
+          secondary: .geminiModel(name: "gemini-flash-lite-latest")
         )
       )
       _ = ai.generativeModelSession(
         model: .hybridModel(
           primary: FirebaseAI.SystemLanguageModel(),
-          secondary: GeminiLanguageModel(name: "gemini-flash-lite-latest")
+          secondary: ai.geminiModel(name: "gemini-flash-lite-latest")
         )
       )
 
       // Initialize a session with a `HybridModel` of cloud models
-      let gemmaModel = ai.geminiLanguageModel(name: "gemma-4-31b-it")
+      let gemmaModel = ai.geminiModel(name: "gemma-4-31b-it")
       let cloudHybridModel = HybridModel(primary: gemmaModel, secondary: geminiLanguageModel)
       _ = ai.generativeModelSession(model: cloudHybridModel)
