+++ FirebaseAI/Tests/Unit/GenerativeModelSessionTests.swift
@@ -195,7 +195,7 @@
       // Verify first request succeeds.
       let response1 = try await session.respond(to: testPrompt)
       XCTAssertEqual(response1.content, "Mountain View")
-      XCTAssertEqual(response1.rawResponse.modelVersion, model1.modelConfig.modelName)
+      XCTAssertEqual(response1.rawResponse.modelVersion, model1.modelName)
 
       // Verify no fallback to model2 after successful request.
       await XCTAssertThrowsError({
@@ -241,7 +241,7 @@
       let response1 = try await session.respond(to: testPrompt)
 
       XCTAssertEqual(response1.content, "Mountain View")
-      XCTAssertEqual(response1.rawResponse.modelVersion, model2.modelConfig.modelName)
+      XCTAssertEqual(response1.rawResponse.modelVersion, model2.modelName)
       XCTAssertTrue(MockURLProtocol.requestHandlersQueue.isEmpty, """
       Expected the queue to be empty after automatically falling back to model2.
       """)
@@ -623,8 +623,8 @@
                          firebaseInfo: FirebaseInfo? = nil, apiConfig: APIConfig? = nil,
                          safetySettings: [SafetySetting]? = nil,
                          requestOptions: RequestOptions? = nil, urlSession: URLSession? = nil)
-      throws -> GeminiLanguageModel {
-      return GeminiLanguageModel(
+      throws -> GeminiModel {
+      return GeminiModel(
         modelName: modelName ?? testModelName,
         modelResourceName: modelResourceName ?? testModelResourceName,
         firebaseInfo: firebaseInfo ?? GenerativeModelTestUtil.testFirebaseInfo(),
